{"schema":"decodex/commit/1","summary":"Extract startup auxiliary window owner","authority":"manual"}
